+++ src/ripe_rainbow/interactive/__init__.py
@@ -2,7 +2,9 @@
 # -*- coding: utf-8 -*-
 
 from . import drivers
+from . import events
 from . import test_cases
 
 from .drivers import InteractiveDriver, SeleniumDriver
+from .events import EVENT_STRINGIFIERS
 from .test_cases import InteractiveTestCase